Refactored plotter splitting to several smaller ones
